Finalized codes

+++ reducer2.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 import sys
 word_dict = {}
 for line in sys.stdin:
@@ -12,5 +13,5 @@ for line in sys.stdin:
     num = line.strip().split("=")[1]
     word = line.strip().split(",")[1].split("=")[0]
     total_word = word_dict[document]
-    frequency = int(num)/total_word
+    frequency = float(num)/total_word
     sys.stdout.write("{0}@{1},{2}\n".format(word, document, frequency))
